requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
dj-rest-auth==5.0.2
Django==5.0
django-extensions==3.2.3
djangorestframework==3.14.0
executing==2.0.1
icecream==2.1.3
Pygments==2.17.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
